Add files via upload
Many mentally ill people avoid going to a hospital for counselling because of the stigma attached to them by 
society. Many problems have been identified such as the lack of an online guidance system for Sri Lankans to 
use.
The solution aims to produce a web application with AI chatbot that provides users to Opportunity to exchange
text messages with counsellors for people who are reluctant to speak, Facility to send a report on the mental 
level of the patient to the counsellor after texting, From that report, the counsellor will have a full 
understanding of the patient at their next meeting with the patient, Facilitate access to services through video 
phone calls from home with the consultant without being noticed by the society, There is also the facility to 
turn off the camera and exchange text messages with the advisor during the video call. Even for a person who 
is unable to speak, this video call can be used to present problems and get solutions through sign language